Adds options to make running preprocessing on the cluster easier
This commit adds two new optional command-line arguments, plugin and
working_dir, to make running the preprocessing pipeline on the cluster
easier. working_dir sets the working directory; if unset, this will be
the same as the output directory and is provided because the working
directory will typically be too large for a user's home directory on
the cluster (it should instead be set to the scratch
directory). plugin is provided to allow the user to choose the run
plugin; if unset, it will be MultiProc, the previous choice here. on
the cluster, it should be SLURM (MultiProc will use up too much memory
on the node and get killed)

also adds a reminder to the subject argument help that the script will
look for this subject in Freesurfer's $SUBJECT_DIR