BIDS info now grabbed from BIDS layout
I realized that, if the data is BIDS-formatted, all the relevant
information can be grabbed from BIDS directly, requiring no user
input. To do this, we use pybids' GrabBids library so I don't have to
implement this myself. But since you don't need pybids if your data is
in prisma format, we wrap that import in a try/except block

 - this means a bunch of arguments (everything except datadir and
   outdir) are no longer required, since they can be inferred if the
   data is BIDS-formatted. this should still throw an exception if your
   running this on prisma-formatted data and don't supply them, but it
   will happen further down, when the script tries to access them from
   the args dict

 - also adds a bit of info about freesurfer and BIDS compatibility in
   the help string there for -dir_structure