Bugfix to handle 3d MGH images
When your input nifti is only 3d (that is, there's only one TR; this
happens with the R2 nifti created by GLMdenoise), the np.expand_dims
call that was being used expanded it to 4 dimensions anyway, which
caused all sorts of issues with how nibabel handled it. with this, we
now will properly only expand the data to 3 dimensions, and so
everything gets handled properly.